examples/web_demo/requirements.txt
Jinja2==2.8
Pillow==2.7.0
cv2==1.0
h5py==2.5.0
html5lib==0.999
httplib2==0.9
lmdb==0.87
matplotlib==1.4.3
mpld3==0.2
numpy==1.9.1
protobuf==2.6.1
pyzmq==14.7.0
scikit-image==0.11.3
scipy==0.16.0
tornado==4.2.1
urllib3==1.9.1
